+++ src/components/__tests__/Details.test.js
@@ -3,8 +3,8 @@ import { render, screen } from '@testing-library/react';
 import renderer from 'react-test-renderer';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
-import Details from './Details';
-import store from './redux/configureStore';
+import Details from '../Details';
+import store from '../redux/configureStore';
 
 const MockDetails = () => (
   <Provider store={store}>
