+++ src/components/__tests__/Home.test.js
@@ -1,10 +1,9 @@
 import { render, screen } from '@testing-library/react';
 import renderer from 'react-test-renderer';
-// import userEvent from '@testing-library/user-event';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
-import store from './redux/configureStore';
-import Home from './Home';
+import Home from '../Home';
+import store from '../redux/configureStore';
 
 const MockHome = () => (
   <Provider store={store}>
@@ -50,16 +49,3 @@ describe('Check if Home Component renders correctly', () => {
     expect(tree).toMatchSnapshot();
   });
 });
-
-// describe('Check if Details Component renders correctly', () => {
-//   test('renders heading', async () => {
-//     jest.mock('./__mocks__/api.js');
-//     render(<MockHome />);
-//     userEvent.click(screen.getByRole('link', {
-//       name: /01 bulbasaur/i,
-//     }));
-//     await expect(screen.getByRole('heading', {
-//       name: /bulbasaur/i,
-//     })).resolves.toBeVisible();
-//   });
-// });
