+++ src/components/__tests__/MainPreview.test.js
@@ -2,8 +2,8 @@ import { render, screen } from '@testing-library/react';
 import renderer from 'react-test-renderer';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
-import MainPreview from './MainPreview';
-import store from './redux/configureStore';
+import MainPreview from '../MainPreview';
+import store from '../redux/configureStore';
 
 const MockMainPreview = () => (
   <Provider store={store}>
