+++ src/components/__tests__/MoreDetails.test.js
@@ -1,8 +1,8 @@
 import renderer from 'react-test-renderer';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
-import MoreDetails from './MoreDetails';
-import store from './redux/configureStore';
+import MoreDetails from '../MoreDetails';
+import store from '../redux/configureStore';
 
 const MockMoreDetails = () => (
   <Provider store={store}>
